refactor: Restaurant 패키지에 dao 적용 (#415)
