+++ backend/src/main/java/com/celuveat/restaurant/query/dao/RestaurantLikeQueryDaoSupport.java
@@ -1,4 +1,4 @@
-package com.celuveat.restaurant.query;
+package com.celuveat.restaurant.query.dao;
 
 import com.celuveat.restaurant.command.domain.Restaurant;
 import com.celuveat.restaurant.command.domain.RestaurantLike;
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface RestaurantLikeQueryRepository extends JpaRepository<RestaurantLike, Long> {
+public interface RestaurantLikeQueryDaoSupport extends JpaRepository<RestaurantLike, Long> {
 
     Optional<RestaurantLike> findByRestaurantAndMemberId(Restaurant restaurant, Long memberId);
